# schema/flow-server.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flow-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/deployment/start": {
      "post": {
        "tags": [
          "deployments"
        ],
        "operationId": "start_deployment_doc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Deployment id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flow",
            "in": "query",
            "description": "Flow id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Deployment tag",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/StartDeploymentParamsDoc"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Start deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowRunTokenOutputDoc"
                }
              }
            }
          }
        }
      }
    },
...
    "/flow/start/{id}": {
      "post": {
        "tags": [
          "flows"
        ],
        "operationId": "start_flow_doc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Flow id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/StartFlowParamsDoc"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Start owned or shared flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowRunStartOutputDoc"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlowInputValueDoc": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/IValueDoc"
          },
          {
            "$ref": "#/components/schemas/JsonValueDoc"
          }
        ]
      },
      "FlowRunStartOutputDoc": {
        "type": "object",
        "required": [
          "flow_run_id"
        ],
        "properties": {
          "flow_run_id": {
            "type": "string"
          }
        }
      },
      "FlowRunTokenOutputDoc": {
        "type": "object",
        "required": [
          "flow_run_id",
          "token"
        ],
        "properties": {
          "flow_run_id": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "IValueDoc": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "S"
            ],
            "properties": {
              "S": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "D"
            ],
            "properties": {
              "D": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "I"
            ],
            "properties": {
              "I": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "U"
            ],
            "properties": {
              "U": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "I1"
            ],
            "properties": {
              "I1": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "U1"
            ],
            "properties": {
              "U1": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "F"
            ],
            "properties": {
              "F": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "B"
            ],
            "properties": {
              "B": {
                "type": "boolean"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "N"
            ],
            "properties": {
              "N": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "B3"
            ],
            "properties": {
              "B3": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "B6"
            ],
            "properties": {
              "B6": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "BY"
            ],
            "properties": {
              "BY": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "A"
            ],
            "properties": {
              "A": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IValueDoc"
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "M"
            ],
            "properties": {
              "M": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/IValueDoc"
                },
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "JsonValueDoc": {
        "oneOf": [
          {
            "type": "null"
          },
          {
            "$ref": "#/components/schemas/JsonValueDocNonNull"
          }
        ]
      },
      "JsonValueDocNonNull": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "number",
            "format": "double"
          },
          {
            "type": "boolean"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonValueDoc"
            }
          },
          {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValueDoc"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        ]
      },
...
      "PartialConfigDoc": {
        "type": "object",
        "required": [
          "only_nodes",
          "values_config"
        ],
        "properties": {
          "only_nodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "values_config": {
            "$ref": "#/components/schemas/ValuesConfigDoc"
          }
        }
      },
...
      "StartDeploymentParamsDoc": {
        "type": "object",
        "properties": {
          "action_signer": {
            "type": [
              "string",
              "null"
            ]
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/FlowInputValueDoc"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "StartFlowParamsDoc": {
        "type": "object",
        "properties": {
          "environment": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "inputs": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/FlowInputValueDoc"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "output_instructions": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "partial_config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartialConfigDoc"
              }
            ]
          }
        }
      },
...
      "ValuesConfigDoc": {
        "type": "object",
        "required": [
          "nodes"
        ],
        "properties": {
          "default_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "nodes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      }
    }
  },
...
}